Cleanup old transaction list files
